chore(module): renamed directories; added module-builder
