+++ build.config.mjs
@@ -1,0 +1,12 @@
+export default {
+  failOnWarn: false,
+  externals: [
+    'consola',
+    'webpack',
+    'change-case',
+    'html-webpack-plugin',
+    'webpack-bundle-analyzer',
+    'clone',
+    'vue'
+  ]
+};
