+++ src/tmpl/entry.mjs
